utils/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "meld-utils"
version = "0.1.0"
description = "Shared utilities for MELD"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["pyyaml"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["meld_utils*"]